+++ lib/gitlab_s.mli
@@ -272,6 +272,7 @@ module type Gitlab = sig
     val get :
       ?fail_handlers:'a parse handler list ->
       ?expected_code:Cohttp.Code.status_code ->
+      ?media_type:string ->
       ?headers:Cohttp.Header.t ->
       ?token:Token.t ->
       ?params:(string * string) list ->
@@ -1307,6 +1308,43 @@ module type Gitlab = sig
       unit ->
       Gitlab_t.runners Response.t Monad.t
     (** [list ~token] Get a list of specific runners available to the user.*)
+  end
+
+  (** The [Job Artifacts] module provides access to {{:https://docs.gitlab.com/ee/api/job_artifacts.html}Job Artifacts API}.
+  *)
+  module Job_artifacts : sig
+    val get :
+      ?token:Token.t ->
+      project_id:int ->
+      job_id:int ->
+      ?job_token:string ->
+      unit ->
+      string option Response.t Monad.t
+    (** [artifacts ~project_id ~job_id ()] Get the job’s artifacts zipped archive of a project. *)
+
+    val get_archive :
+      ?token:Token.t ->
+      project_id:int ->
+      ref_name:string ->
+      job:string ->
+      ?job_token:string ->
+      unit ->
+      string option Response.t Monad.t
+    (** [archive ~project-id ~ref_name ~job] Download the artifacts zipped archive from
+        the latest successful pipeline for the given reference name [ref_name]
+        and [job], provided the job finished successfully.  *)
+
+    val get_file :
+      ?token:Token.t ->
+      project_id:int ->
+      job_id:int ->
+      artifact_path:string ->
+      ?job_token:string ->
+      unit ->
+      string option Response.t Monad.t
+    (** [archive ~project-id ~ref_name ~job] Download a single
+        artifact file from a [job_id] from inside the job’s artifacts
+        zipped archive.  *)
   end
 end